+++ lib/dune/parser/unsafe_ast.ex
@@ -3,8 +3,9 @@ defmodule Dune.Parser.UnsafeAst do
 
   @type t :: %__MODULE__{
           ast: Macro.t(),
-          atom_mapping: Dune.AtomMapping.t()
+          atom_mapping: Dune.AtomMapping.t(),
+          stdio: binary()
         }
   @enforce_keys [:ast, :atom_mapping]
-  defstruct @enforce_keys
+  defstruct @enforce_keys ++ [stdio: <<>>]
 end