+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - `Dune.Allowlist.Default` allows the new `Duration` module and new kernel
   functions
 - Add an `:inspect_sort_maps` option for deterministic outputs
+- Capture and return parser warnings in `stdio`
 
 ## v0.3.6 (2023-12-23)